Can now remove images
Images can be removed now. Remove on breadcrumbs does not work yet. It's
a problem in the way I set up the fragment and adapter. I did it this
way because currently we are displaying the crumbs in a listview. Since
we are updating this to a google maps image, I put all the functionality
into the fragment instead of the adapter, thus cannot easily delete
crumbs from the app. Will easily be fixed when google map interface is
complete