final push
i added some extra features like choosing when to save images to
gallery.  also redid the breadcrumbs buttons.  There is a small error.
I think we need to do something to our databse references in each
fragment so they stop listening when that fragment is not up or at least
add if statements to prevent double listening